+++ package.json
@@ -37,6 +37,7 @@
         "@typescript-eslint/eslint-plugin": "^3.0.1",
         "@typescript-eslint/parser": "^3.0.1",
         "babel-loader": "^8.1.0",
+        "babel-plugin-emotion": "^10.0.33",
         "chromatic": "^5.0.0",
         "emotion": "^10.0.27",
         "eslint": "^7.1.0",
@@ -77,8 +78,8 @@
     "peerDependencies": {
         "@guardian/src-foundations": "^1.1.0",
         "@guardian/src-icons": "^1.9.1",
-        "@emotion/core": "^10.0.28",
         "emotion": "^10.0.27",
+        "babel-plugin-emotion": "^10.0.33",
         "react": "^16.13.1",
         "react-dom": "^16.13.1"
     },
